# GS-68:  Remove Activity Date Filter