--- a/css/style.css
+++ b/css/style.css
@@ -78,6 +78,6 @@ select.relativeFilter {
   vertical-align: middle;
 }
 
-#pivot-report-config .hidden {
+#pivot-report-config .hidden, #pivot-report-filters {
   display: none;
 }